requirements.txt
# SEO Intelligence Hub — Python Dependencies
# Install with: pip install -r requirements.txt

# Core LLM & Agent
langchain>=0.2.0
langchain-anthropic>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
langsmith>=0.1.0

# Web Scraping & SERP
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Embeddings & Clustering
openai>=1.30.0
scikit-learn>=1.3.0
hdbscan>=0.8.0

# UI (optional — for Gradio demo)
gradio>=4.0.0

# Env & Config
python-dotenv>=1.0.0

# Utilities
tqdm>=4.66.0
